xtensa: deliver signal raised in double exception handler
This fixes infinite loop caused by an exception in a fast exception
handler, e.g. on userspace access to unaligned address in unmapped memory
region, or spill_registers syscall when user stack pointer is corrupted.

In such situations the handler of the nested exception sends a signal that
is either handled by the task or that kills the task avoiding infinite
retry of the faulty operation.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>